dashboard/auth_portal_4430/requirements.txt
flask==3.0.0
python3-saml==1.15.0
PyJWT==2.8.0
redis==5.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
cryptography==41.0.7
lxml==4.9.3
xmlsec==1.3.13
gunicorn==21.2.0
# OIDC support
authlib==1.3.0
httpx==0.27.0
